size_t not int

+++ src/fmalloc.c
@@ -9,14 +9,14 @@ extern "C" {
 #endif
 
 static void
-fail_die_malloc_die(int size, const char *file, int line)
+fail_die_malloc_die(size_t size, const char *file, int line)
 {
   fprintf(stderr, "Out of memory! (wanted %d at %s:%d)\n", size, file, line);
   exit(1);
 }
 
 void *
-fail_die_malloc(int size, const char *file, int line)
+fail_die_malloc(size_t size, const char *file, int line)
 {
   void *p = malloc(size);
   if (!p && size)
@@ -25,7 +25,7 @@ fail_die_malloc(int size, const char *file, int line)
 }
 
 void *
-fail_die_realloc(void *p, int size, const char *file, int line)
+fail_die_realloc(void *p, size_t size, const char *file, int line)
 {
   if (!p)
     return fail_die_malloc(size, file, line);
